+++ src/array/drop.js
@@ -15,6 +15,8 @@ import _curry2 from '../_internals/_curry2.js'
  * import { drop } from 'kyanite'
  *
  * drop(3, [1, 2, 3, 4, 5]) // => [4, 5]
+ * drop(6, [1, 2, 3, 4, 5]) // => []
+ * drop(-1, [1, 2, 3, 4, 5]) // => [1, 2, 3, 4, 5]
  *
  * // It's also curried
  *
@@ -22,6 +24,6 @@ import _curry2 from '../_internals/_curry2.js'
  *
  * d([1, 2, 3, 4, 5]) // => [4, 5]
  */
-const drop = (i, list) => list.slice(i, Infinity)
+const drop = (i, list) => list.slice(Math.max(0, i), Infinity)
 
 export default _curry2(drop)