+++ tests/array/drop.js
@@ -8,10 +8,24 @@ test('drop -- Handles dropping values', t => {
   t.end()
 })
 
+test('drop -- Handles dropping values when number is larger than length', t => {
+  const results = drop(6, [1, 2, 3, 4, 5])
+
+  t.same(results, [])
+  t.end()
+})
+
 test('drop -- Handles invalid data sets', t => {
   const results = drop(3, [])
 
   t.same(results, [])
+  t.end()
+})
+
+test('drop -- Handles negative index', t => {
+  const results = drop(-1, [1, 2, 3, 4, 5])
+
+  t.same(results, [1, 2, 3, 4, 5])
   t.end()
 })