+++ types/index.d.ts
@@ -261,6 +261,12 @@ declare namespace K {
     drop<T = any>(n: number): (list: ReadonlyArray<T> | string) => T[] | string;
 
     /**
+     * Returns a list containing all but the last n elements of the given list.
+     */
+    dropLast<T = any>(n: number, list: ReadonlyArray<T> | string): T[] | string;
+    dropLast<T = any>(n: number): (list: ReadonlyArray<T> | string) => T[] | string;
+
+    /**
      * Runs through an array and drops values so long as the function used returns true once it returns false the iteration will stop
      */
     dropWhile<T = any>(fn: (a: T) => boolean, arr: ReadonlyArray<T> | string): T[] | string;
@@ -654,9 +660,9 @@ declare namespace K {
     /**
      * A safe way to find an item within an object, will return the provided default if it's not found or the value itself if it is found
      */
-    pathOr<T = any>(a: any, keys: Path, obj: object): T;
-    pathOr(a: any, keys: Path): <T = any>(obj: object) => T;
-    pathOr(a: any): (keys: Path) => <T = any>(obj: object) => T;
+    pathOr<T = any, U = any>(a: U, keys: Path, obj: object): T;
+    pathOr<U = any>(a: U, keys: Path): <T = any>(obj: object) => T;
+    pathOr<U = any>(a: U): (keys: Path) => <T = any>(obj: object) => T;
 
     /**
      * Runs a path check on a given object and then runs a predicate function on the result
@@ -664,6 +670,12 @@ declare namespace K {
     pathSatisfies<T = any, U = any>(pred: (val: T) => boolean, keys: Path, obj: U): boolean;
     pathSatisfies<T = any, U = any>(pred: (val: T) => boolean, keys: Path): (obj: U) => boolean;
     pathSatisfies<T = any, U = any>(pred: (val: T) => boolean): (keys: Path) => (obj: U) => boolean;
+
+    /**
+     * Picks only the requested keys from a provided object
+     */
+    pick<U = any>(keys: ReadonlyArray<string>, obj: U): U;
+    pick<U = any>(keys: ReadonlyArray<string>): (obj: U) => U;
 
     /**
      * Applies a sequence of transformations over a value
@@ -891,6 +903,12 @@ declare namespace K {
      */
     take<T = any>(i: number, list: ReadonlyArray<T> | string): T[] | string;
     take<T = any>(i: number): (list: ReadonlyArray<T> | string) => T[] | string;
+
+    /**
+     * Returns a new list containing the last n elements of the given list. If n > list.length, returns a list of list.length elements.
+     */
+    takeLast<T = any>(n: number, list: ReadonlyArray<T> | string): T[] | string;
+    takeLast<T = any>(n: number): (list: ReadonlyArray<T> | string) => T[] | string;
 
     /**
      * Takes values from an array so long as the predicate function continues to return true